+++ android/ijkplayer/ijkplayer-example/src/main/java/tv/danmaku/ijk/media/example/application/IjkApp.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright (C) 2020 Befovy
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package tv.danmaku.ijk.media.example.application;
+
+import android.app.Application;
+import android.os.StrictMode;
+
+import com.github.anrwatchdog.ANRWatchDog;
+
+import tv.danmaku.ijk.media.example.BuildConfig;
+
+public class IjkApp extends Application {
+
+    @Override
+    public void onCreate() {
+        super.onCreate();
+        if (BuildConfig.DEBUG) {
+            StrictMode.setThreadPolicy(new StrictMode.ThreadPolicy.Builder()
+                    .detectAll()
+                    .penaltyLog()
+                    .build());
+            new ANRWatchDog().setIgnoreDebugger(true).start();
+        }
+    }
+}