+++ ijkmedia/ijksdl/ijksdl_egl.c
@@ -32,6 +32,7 @@
 
 #if ANDROID
 #include <android/native_window.h>
+#include "ijksdl/android/ijksdl_android_jni.h"
 #endif
 
 #define IJK_EGL_RENDER_BUFFER 0
@@ -276,6 +277,10 @@ static EGLBoolean IJK_EGL_prepareRenderer(IJK_EGL* egl, SDL_VoutOverlay *overlay
 
     IJK_EGL_Opaque *opaque = egl->opaque;
 
+    if (egl->amc_surface_changed) {
+        IJK_GLES2_Renderer_AMC_set_texture(opaque->renderer, egl->amc_surface);
+        egl->amc_surface_changed = 0;
+    }
     if (!IJK_GLES2_Renderer_isValid(opaque->renderer) ||
         !IJK_GLES2_Renderer_isFormat(opaque->renderer, overlay->format)) {
 
@@ -286,6 +291,10 @@ static EGLBoolean IJK_EGL_prepareRenderer(IJK_EGL* egl, SDL_VoutOverlay *overlay
         if (!opaque->renderer) {
             ALOGE("[EGL] Could not create render.");
             return EGL_FALSE;
+        }
+
+        if (overlay->format == SDL_FCC__AMC) {
+            IJK_GLES2_Renderer_AMC_set_texture(opaque->renderer, egl->amc_surface);
         }
 
         if (!IJK_GLES2_Renderer_use(opaque->renderer)) {
@@ -350,12 +359,29 @@ void IJK_EGL_releaseWindow(IJK_EGL* egl)
     IJK_EGL_terminate(egl);
 }
 
+IJK_GLES2_Renderer *IJK_EGL_get_renderer(IJK_EGL *egl)
+{
+    if (!egl || !egl->opaque)
+        return NULL;
+    return egl->opaque->renderer;
+}
+
 void IJK_EGL_free(IJK_EGL *egl)
 {
     if (!egl)
         return;
 
     IJK_EGL_terminate(egl);
+#if ANDROID
+    if (egl->amc_surface != NULL) {
+        JNIEnv *env = NULL;
+        if (JNI_OK != SDL_JNI_SetupThreadEnv(&env)) {
+            ALOGE("SDL_Android_GetApiLevel: SetupThreadEnv failed");
+            return;
+        }
+        J4A_DeleteGlobalRef(env, egl->amc_surface);
+    }
+#endif
 
     memset(egl, 0, sizeof(IJK_EGL));
     free(egl);