+++ ios/IJKMediaPlayer/IJKMediaPlayer/IJKFFMediaPlayer.h
@@ -43,6 +43,12 @@ typedef NS_ENUM(NSInteger, IJKMPEventType) {
     IJKMPET_PLAYBACK_STATE_CHANGED  = 700,
 };
 
+typedef NS_ENUM(NSInteger, IJKMPErrorCode){
+    IJKMPEC_SNAPSHOT                = -480,
+};
+
+
+typedef void(^OnSnapshotBlock) (UIImage* __nullable image , NSError* __nullable error );
 
 @class IJKFFMediaPlayer;
 
@@ -55,6 +61,9 @@ typedef NS_ENUM(NSInteger, IJKMPEventType) {
 
 @interface IJKFFMediaPlayer : NSObject<IJKCVPBViewProtocol, IJKSDLGLViewProtocol>
 
+
+@property (nonatomic) BOOL ignoreAudioInterrupt;
+
 - (instancetype)init;
 - (instancetype)initWithFbo;
 - (int) setDataSource:(NSString *)url;
@@ -66,6 +75,7 @@ typedef NS_ENUM(NSInteger, IJKMPEventType) {
 - (BOOL) isPlaying;
 - (void) shutdown;
 
+- (void) takeSnapshot:(OnSnapshotBlock) block;
 - (long) getCurrentPosition;
 - (long) getDuration;
 - (int)  seekTo:(long) msec;